Module amazon_sp_vendor_direct_fulfillment_shipping::apis::vendor_shipping_api

Enums

GetPackingSlipError

struct for typed errors of method get_packing_slip

GetPackingSlipsError

struct for typed errors of method get_packing_slips

SubmitShipmentConfirmationsError

struct for typed errors of method submit_shipment_confirmations

SubmitShipmentStatusUpdatesError

struct for typed errors of method submit_shipment_status_updates

Functions

get_packing_slip

Returns a packing slip based on the purchaseOrderNumber that you specify. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_packing_slips

Returns a list of packing slips for the purchase orders that match the criteria specified. Date range to search must not be more than 7 days. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

submit_shipment_confirmations

Submits one or more shipment confirmations for vendor orders. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

submit_shipment_status_updates

This API call is only to be used by Vendor-Own-Carrier (VOC) vendors. Calling this API will submit a shipment status update for the package that a vendor has shipped. It will provide the Amazon customer visibility on their order, when the package is outside of Amazon Network visibility. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.