Module amazon_sp_vendor_direct_fulfillment_shipping::apis::vendor_shipping_labels_api

Enums

GetShippingLabelError

struct for typed errors of method get_shipping_label

GetShippingLabelsError

struct for typed errors of method get_shipping_labels

SubmitShippingLabelRequestError

struct for typed errors of method submit_shipping_label_request

Functions

get_shipping_label

Returns a shipping label for the purchaseOrderNumber that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 10 | 10 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values then those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_shipping_labels

Returns a list of shipping labels created during the time frame that you specify. You define that time frame using the createdAfter and createdBefore parameters. You must use both of these parameters. The date range to search must not be more than 7 days. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 10 | 10 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values then those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

submit_shipping_label_request

Creates a shipping label for a purchase order and returns a transactionId for reference. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 10 | 10 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values then those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.