Function amazon_sp_vendor_direct_fulfillment_shipping::apis::vendor_shipping_labels_api::get_shipping_label

pub async fn get_shipping_label(
    configuration: &Configuration,
    purchase_order_number: &str
) -> Result<GetShippingLabelResponse, Error<GetShippingLabelError>>

Returns a shipping label for the purchaseOrderNumber that you specify. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.