Function amazon_sp_vendor_direct_fulfillment_shipping::apis::vendor_shipping_labels_api::get_shipping_labels

pub async fn get_shipping_labels(
    configuration: &Configuration, 
    created_after: String, 
    created_before: String, 
    ship_from_party_id: Option<&str>, 
    limit: Option<i32>, 
    sort_order: Option<&str>, 
    next_token: Option<&str>
) -> Result<ShippingLabelList, Error<GetShippingLabelsError>>

Returns a list of shipping labels created during the time frame that you specify. You define that time frame using the createdAfter and createdBefore parameters. You must use both of these parameters. The date range to search must not be more than 7 days. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 10 | 10 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values then those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.